+++ Monster.hpp
@@ -6,19 +6,19 @@
 class Monster{
 	public:
 		// note from Harrison: please use camelCase for function and variable names (first letter undercase) as per naming conventions.
-		void setAttack(double val){};
-		void setPhysicalDef(double val);
-		void setMagicalDef(double val);
-		void setMaxHealth(double val);
-		double getAttack();
-		double getMagicalDef();
-		double getPhysicalDef();
-		virtual double monsterHit() = 0;
+		void setAttack(int val){};
+		void setPhysicalDef(int val);
+		void setMagicalDef(int val);
+		void setMaxHealth(int val);
+		int getAttack();
+		int getMagicalDef();
+		int getPhysicalDef();
+		virtual int monsterHit() = 0;
 	private:
-		double maxHealth;
-		double attack;
-		double physicalDef;
-		double magicalDef;
+		int maxHealth;
+		int attack;
+		int physicalDef;
+		int magicalDef;
 };
 
 #endif
